Extract 0-size tensor tests from test_cdist.py into separate test file
Move the newly added zero-size tensor test cases (TestCdistZeroSizeR2,
TestCdistZeroSizeR1, TestCdistZeroSizeC1, TestCdistZeroSizeBatch3D,
TestCdistZeroSizeBatch4D, TestCdistZeroSizeGrad) into a dedicated
test_cdist_zero_size.py file for better test organization.
